1.1.6
Added:
Management(Flight creation and edititing)

Updated:
ALogic
ConstructionYard
Airport
Plane

Airports.xhtml
hangar.xhtml

Notes:
new features involving Flights, and rudimentary work on Removal of Planes in the Hangar.java

+++ src/java/com/Eagle/Admin/ALogic.java
@@ -2,7 +2,9 @@ package com.Eagle.Admin;
 
 //Admin Logic
 import com.Eagle.Model.Airport;
+import com.Eagle.Model.Flight;
 import com.Eagle.Model.Plane;
+import java.util.ArrayList;
 import java.util.List;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
@@ -16,36 +18,63 @@ public class ALogic
     @PersistenceContext
     EntityManager em;
 
+    List<Plane> allPlanes;
+
     public ALogic()
     {
+        allPlanes = new ArrayList<>();
+
     }
 
+    //<editor-fold defaultstate="collapsed" desc="Hangar">
     public void persistAirplane(Plane plane)
     {
         em.persist(plane);
     }
-    
-    
-    public List<Plane> getAllPlanes() 
+
+    public Plane getOnePlane(Long id)
     {
-        Query query = em.createNamedQuery("SELECT * FROM plane", Plane.class);
-        List<Plane> list = query.getResultList();
-        return list;
+        return em.find(Plane.class, id);
+    }
+
+    public List<Plane> getAllPlanes()
+    {
+        Query query = em.createNativeQuery("SELECT * FROM plane", Plane.class);
+        allPlanes = query.getResultList();
+        return allPlanes;
     }
 
     public void removePlane(Plane plane)
     {
-        em.remove(plane);
+        if (plane != null)
+        {
+            em.remove(plane);
+        }
     }
+//</editor-fold>
 
+    //<editor-fold defaultstate="collapsed" desc="Construction Yard">
     public void persistAirport(Airport airport)
     {
         em.persist(airport);
     }
-    
+
     public void removeAirport(Airport airport)
     {
         em.remove(airport);
     }
+//</editor-fold>
+
+    //<editor-fold defaultstate="collapsed" desc="Management">
+    public void persistFlight(Flight flight)
+    {
+        em.persist(flight);
+    }
+
+    public void removeFlight(Flight flight)
+    {
+        em.remove(flight);
+    }
+//</editor-fold>
 
 }
